backend/antopic/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
argilla==1.6.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
auth0-python==4.1.0
backoff==2.2.1
billiard==3.6.4.0
Brotli==1.0.9
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
cryptography==40.0.1
dataclasses-json==0.5.7
Deprecated==1.2.13
Django==4.1.8
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
et-xmlfile==1.1.0
frozenlist==1.3.3
gptcache==0.1.10
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
kombu==5.2.4
langchain==0.0.139
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.1
pipreqs==0.4.13
prompt-toolkit==3.0.38
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.15.0
PyJWT==2.6.0
pymongo==3.12.3
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
sqlparse==0.2.4
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.12
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.2
wrapt==1.14.1
XlsxWriter==3.1.0
yarg==0.1.9
yarl==1.8.2
yt-dlp==2023.3.4
